Add files via upload
Write a program in Python where you generate a random integer from 1 to 15 inclusive where your only random number generator is a six-sided die. Assume a function named roll6 which returns a 1, 2, 3, 4, 5 or 6 and write a function named roll15 which uses the roll6 function and returns an integer between 1 and 15 inclusive, which has an equal probability of each number. Do not use a random number generator other than roll6.